requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
et_xmlfile==2.0.0
idna==3.10
image==1.5.33
openpyxl==3.1.5
pillow==11.0.0
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
